+++ sherlog/__init__.py
@@ -1,0 +1,8 @@
+# coding=utf-8
+
+from sherlog.logger import get_logger
+from sherlog.logger import set_logger
+
+from sherlog.version import VERSION
+
+__version__ = VERSION