+++ union/__init__.py
@@ -1,8 +1,0 @@
-# coding=utf-8
-
-from union.logger import get_logger
-from union.logger import set_logger
-
-from union.version import VERSION
-
-__version__ = VERSION